cid-budget/backend/app/core.py
import os
import logging
import sys
from datetime import datetime, timezone

LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO").upper()
JWT_SECRET = os.getenv("JWT_SECRET", "")
JWT_ALGORITHM = "HS256"
DATABASE_URL = os.getenv("DATABASE_URL", "postgresql://postgres:postgres@localhost:5432/cid_budget")
CORS_ORIGINS = os.getenv("CORS_ORIGINS", "*").split(",")
RATE_LIMIT_PER_MINUTE = int(os.getenv("RATE_LIMIT_PER_MINUTE", "60"))


def setup_logging() -> logging.Logger:
    handler = logging.StreamHandler(sys.stdout)
    handler.setFormatter(logging.Formatter(
        "%(asctime)s [%(levelname)s] %(name)s: %(message)s",
        datefmt="%Y-%m-%dT%H:%M:%S%z",
    ))
    root = logging.getLogger()
    root.setLevel(LOG_LEVEL)
    root.handlers.clear()
    root.addHandler(handler)
    return root


def verify_token(token: str, secret: str = "") -> dict:
    import jwt as pyjwt
    jwt_secret = secret or os.getenv("JWT_SECRET", "")
    if not jwt_secret:
        return {"sub": "anonymous", "role": "admin"}
    try:
        payload = pyjwt.decode(token, jwt_secret, algorithms=[JWT_ALGORITHM])
        return payload
    except pyjwt.ExpiredSignatureError:
        raise PermissionError("Token expired")
    except pyjwt.InvalidTokenError:
        raise PermissionError("Invalid token")
